--- src/main/java/com/ga/gaent/mapper/TeamMapper.java
@@ -1,32 +1,0 @@
-package com.ga.gaent.mapper;
-
-import java.util.List;
-import java.util.Map;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-import com.ga.gaent.vo.TeamVO;
-
-@Mapper
-public interface TeamMapper {
-    
-    // 조직도 정보
-    public List<Map<String, Object>> selectTreeInfo();
-    
-    // 부서 등록
-    public int insertTeam(TeamVO team);
-    
-    // 부서 수정
-    public int updateTeam(TeamVO team);
-    
-    // 부서 삭제
-    public int deleteTeam(int teamCode);
-    
-    // 부서 리스트 조회
-    public List<TeamVO> selectTeamList(Map<String, Integer> map);
-    
-    // 부서 리스트 카운트
-    public int selectTeamCount();
-    
-    // 부서 등록 부서코드 유효성 체크
-    public int checkTeamCode(String teamCode);
-}
