+++ src/main/java/com/ga/gaent/service/EmpService.java
@@ -1,0 +1,5 @@
+package com.ga.gaent.service;
+
+public class EmpService {
+
+}
